pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.3.3",
    "httpx>=0.28.1",
    "pyobjc>=11.0",
    "pyobjc-framework-corelocation>=11.0",
    "requests>=2.32.3",
]

[project.scripts]
location = "location:main"
weather = "weather:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
